# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [projects]
      summary: List tasks for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [projects]
      summary: Create a task
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /projects/{projectId}/tasks/{taskId}:
    patch:
      operationId: patchTask
      tags: [projects]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [projects]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/tasks/suggest:
    post:
      operationId: suggestTasks
      tags: [projects]
      summary: AI-generated task suggestions for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Suggested tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSuggestions"
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        assigneeId:
          type: integer
          nullable: true
        assigneeEmail:
          type: string
          nullable: true
        createdBy:
          type: integer
        dueDate:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, title, status, createdBy, createdAt, updatedAt]
    NewTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        dueDate:
          type: string
      required: [title]
    TaskPatch:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        assigneeId:
          type: integer
          nullable: true
        dueDate:
          type: string
          nullable: true
    TaskSuggestion:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required: [title, description]
    TaskSuggestions:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/TaskSuggestion"
      required: [suggestions]
...