# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  projectId: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  assigneeId?: number | null;
  assigneeEmail?: string | null;
  createdBy: number;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}