# lib/api-zod/src/generated/types/taskPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPatchStatus } from './taskPatchStatus';

export interface TaskPatch {
  title?: string;
  description?: string | null;
  status?: TaskPatchStatus;
  assigneeId?: number | null;
  dueDate?: string | null;
}